图片上传，合成二维码

+++ app/views.py
@@ -7,6 +7,7 @@ import json
 import sqlalchemy.exc
 from tools import Tools
 from werkzeug import secure_filename
+from datetime import datetime
 upload_path = app.config["UPLOAD_FOLDER"]
 
 
@@ -26,25 +27,28 @@ def allowed_file(filename):
 @app.route('/user<int:user_id>_edit/',methods=['GET', 'POST'])
 def electronic_edit(user_id):
     if request.method == 'POST':
+        print "electronic_edit POST"
         user = User.getUser(user_id)
-        if os.path.exists(user.logo):
+        if user.logo and os.path.exists(user.logo):
             os.remove(user.logo)
-        if os.path.exists(user.headpic):
+        if user.headpic and os.path.exists(user.headpic):
             os.remove(user.headpic)
         file = request.files['file']
         logo = request.files["logo"]
+        time = str(datetime.today()).replace(" ","_").replace(":","_").replace(".","_")
         if file and allowed_file(file.filename):
             filename = secure_filename(file.filename)
-            file_path = os.path.join(upload_path, "user_id_"+str(user_id)+"_"+filename)
-            user.headpic = "user_id_"+str(user_id)+"_"+filename
+            file_path = os.path.join(upload_path, "user_id_"+str(user_id)+"_"+time+filename)
+            user.headpic = file_path
             file.save(file_path)
             db.session.commit()
         if logo and allowed_file(logo.filename):
             filename = secure_filename(logo.filename)
-            logo_path = os.path.join(upload_path, "user_id_"+str(user_id)+"_"+filename)
-            user.logo = "user_id_"+str(user_id)+"_"+filename
+            logo_path = os.path.join(upload_path, "user_id_"+str(user_id)+"_"+time+filename)
+            user.logo = logo_path
             logo.save(logo_path)
             db.session.commit()
+        user.makeQrcode(user,user.headpic)#生成二维码
         return redirect(url_for("electronic_finish",user_id=user_id))
 
     return render_template("electronic_edit.html",user_id=user_id)
@@ -52,6 +56,7 @@ def electronic_edit(user_id):
 #保存名片信息到数据库
 @app.route('/_saveContact/',methods=["GET","POST"])
 def saveContact():
+    print "saveContact"#首先进入
     user_id = int(request.form['id'])
     info = json.loads(request.form["info"])
     intro = json.loads(request.form["intro"])
@@ -70,7 +75,8 @@ def saveContact():
     for i in custom:
         c = Contact(_type = i["type"] ,title = i["customTitle"] , text = i["customText"],user_id=user_id )
         db.session.add(c)
-    user.makeQrcode(user)#生成二维码
+
+
     db.session.commit()
     return redirect(url_for("electronic_finish",user_id=user_id))
 
@@ -80,7 +86,19 @@ def saveContact():
 @app.route('/user<int:user_id>_finish/')
 def electronic_finish(user_id):
     user = User.getUser(user_id)
-    return render_template("electronic_finish.html",user_id=user_id,user=user)
+    qrcode = os.path.basename(user.qrcode)
+    if user.logo :
+        logo = os.path.basename(user.logo)
+    else:
+        print "using default logo"
+        logo = "default_logo.png"
+    if user.headpic :
+        headpic = os.path.basename(user.headpic)
+    else:
+        print "using default head"
+        headpic = "default_headpic.png"
+
+    return render_template("electronic_finish.html",user_id=user_id,qrcode=qrcode,logo=logo,headpic=headpic)
 
 @app.route('/user<int:user_id>_card/')
 def card(user_id):
@@ -90,3 +108,9 @@ def card(user_id):
 def _getCardDetail():
     user_id = request.args.get("user_id")
     return Tools.getcarddetail(user_id)
+
+@app.route('/user<int:user_id>_editcard/')
+def editcard(user_id):
+    return render_template("edit-fang.html",user_id=user_id)
+
+
